+++ src/dom/index.ts
@@ -1,3 +1,4 @@
+export { Attribute } from './attribute';
 export { DOMNode, NodeClosed } from './domnode';
 export { DOMTokenList } from './domtokenlist';
 export { DOMTree } from './domtree';